imagetoimage/requirements.txt
# Flask Web框架
Flask==3.0.0

# OpenAI客户端库（用于调用千帆API）
openai==1.3.7

# Flask依赖
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# HTTP请求相关
requests==2.31.0
urllib3==2.0.7

# 环境变量管理
python-dotenv==1.0.0

# 测试相关
pytest==7.4.3
pytest-flask==1.3.0 